Simplifies target device support (#16)
* Configures app for portrait orientation.

Sets the app to only support portrait orientation on all devices.
Adds description for location usage.
Declares that the app does not use non-exempt encryption.

* Simplifies target device family setting

Removes iPad as a supported device family.
This simplifies the project settings and ensures the app is only targeted for iPhone.

Updates the localization to a specific Spanish locale (es-ES).